numerous fixes

+++ content.php
@@ -67,6 +67,7 @@
 						echo "\n\t\t<li><a id='lesson".$i."-".$j."'class='hvr-curl-bottom-right enabledLink' href='?p=".$i."&pp=".$j."' >".trim($submenu[$i][$j])."</a></li>";
 					/*echo " <li><a  id='submenuLink".$page."-".$i."' class=".($i==$subpage && $subpage!="test" ?"'active'":"'enabledSubmenuLink'")." href='?p=".$page."&pp=".$i."'>".$submenu[$i]."</a></li>";
 					*/}
+					if( file_exists ("strony/".$i."/endboss.html"))
 					echo "<li><a id='test".$i."' href='?p=".$i."&pp=test' class='hvr-curl-bottom-right enabledLink".($subpage=="test" ?"'active'":"''")."' >Starcie z Konfuzjuszem</a></li>";
 					echo "\n\t</ul></li>";
 				}
